mlops-structure/requirements.txt
arrow==1.1.1
binaryornot==0.4.4
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
cookiecutter==1.7.3
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jinja2-time==0.2.0
joblib==1.0.1
MarkupSafe==2.0.1
nltk==3.6.2
poyo==0.5.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-slugify==5.0.2
regex==2021.7.6
requests==2.26.0
six==1.16.0
text-unidecode==1.3
textblob==0.15.3
tqdm==4.62.0
urllib3==1.26.6
Werkzeug==2.0.1
